Refactor OverlayMenu into App component
