+++ src/containers/overlayMenu/overlayMenu.js
@@ -1,6 +1,6 @@
 import React, { forwardRef } from 'react';
 import './overlayMenu.css';
-import { useNavigationHelpers } from '../../../utilities/navigationHelpers';
+import { useNavigationHelpers } from '../../utilities/navigationHelpers';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faArrowLeft } from '@fortawesome/free-solid-svg-icons';
